[flphone] update FLNX to the 1.1.8
git-svn-id: https://linux-on-sx1.svn.sourceforge.net/svnroot/linux-on-sx1/trunk/target/src/flphone@152 ca9450b1-263c-0410-a22d-a1a5e2d30397

+++ libs/flnx/test/cursor.cxx
@@ -1,9 +1,9 @@
 //
-// "$Id: cursor.cxx,v 1.1.1.1 2003/08/07 21:18:42 jasonk Exp $"
+// "$Id: cursor.cxx 5519 2006-10-11 03:12:15Z mike $"
 //
 // Cursor test program for the Fast Light Tool Kit (FLTK).
 //
-// Copyright 1998-1999 by Bill Spitzak and others.
+// Copyright 1998-2005 by Bill Spitzak and others.
 //
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Library General Public
@@ -20,7 +20,9 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
 // USA.
 //
-// Please report all bugs and problems to "fltk-bugs@easysw.com".
+// Please report all bugs and problems on the following page:
+//
+//     http://www.fltk.org/str.php
 //
 
 #include <FL/Fl.H>
@@ -37,7 +39,7 @@ Fl_Cursor cursor = FL_CURSOR_DEFAULT;
 Fl_Hor_Value_Slider *cursor_slider;
 
 void choice_cb(Fl_Widget *, void *v) {
-  cursor = (Fl_Cursor)(int)v;
+  cursor = (Fl_Cursor)(long)v;
   cursor_slider->value(cursor);
   fl_cursor(cursor,fg,bg);
 }
@@ -159,5 +161,5 @@ int main(int argc, char **argv) {
 }
 
 //
-// End of "$Id: cursor.cxx,v 1.1.1.1 2003/08/07 21:18:42 jasonk Exp $".
+// End of "$Id: cursor.cxx 5519 2006-10-11 03:12:15Z mike $".
 //
